+++ lib/ChatType.ts
